fnly

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ykai-yua <ykai-yua@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/23 15:34:05 by ykai-yua          #+#    #+#             */
-/*   Updated: 2024/02/24 23:13:57 by ykai-yua         ###   ########.fr       */
+/*   Updated: 2024/08/08 19:20:32 by ykai-yua         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
